+++ src/app/app.module.ts
@@ -12,6 +12,9 @@ import {HomeModule} from "./features/home/home.module";
 import {PropertyModule} from "./features/property/property.module";
 import {OwnerModule} from "./features/owner/owner.module";
 import {TenantModule} from "./features/tenant/tenant.module";
+import {DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE} from "@angular/material/core";
+import {MAT_MOMENT_DATE_FORMATS, MomentDateAdapter} from "@angular/material-moment-adapter";
+import 'moment/locale/fr';
 
 @NgModule({
   declarations: [
@@ -34,6 +37,9 @@ import {TenantModule} from "./features/tenant/tenant.module";
     {
       provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true
     },
+    {provide: MAT_DATE_LOCALE, useValue: 'fr-FR'},
+    {provide: DateAdapter, useClass: MomentDateAdapter, deps: [MAT_DATE_LOCALE]},
+    {provide: MAT_DATE_FORMATS, useValue: MAT_MOMENT_DATE_FORMATS},
   ],
   bootstrap: [AppComponent]
 })