+++ src/app/common/business-models/client.business-model.ts
@@ -1,8 +1,10 @@
+import {AddressBusinessModel} from "./address.business-model";
+
 export interface ClientBusinessModel {
   firstName: string;
   lastName: string;
   birthday: Date;
   phoneNumber: string;
   email: string;
-  address: string [];
+  address: AddressBusinessModel;
 }