+++ src/app/features/tenant/create-tenant/create-tenant.component.ts
@@ -1,10 +1,56 @@
-import { Component } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
+import {FormBuilder, Validators} from "@angular/forms";
+import {BreakpointObserver} from "@angular/cdk/layout";
+import {map, Observable} from "rxjs";
+import {StepperOrientation} from "@angular/material/stepper";
+import * as moment from 'moment';
 
 @Component({
   selector: 'app-create-tenant',
   templateUrl: './create-tenant.component.html',
   styleUrls: ['./create-tenant.component.scss']
 })
-export class CreateTenantComponent {
+export class CreateTenantComponent implements OnInit {
+  identityFormGroup = this._formBuilder.group({
+    firstName: ['', Validators.required],
+    lastName: ['', Validators.required],
+    birthday: [''],
+  });
+  postalAddressFormGroup = this._formBuilder.group({
+    road: ['', Validators.required],
+    additionalAddress: [''],
+    town: ['', Validators.required],
+    // Matches 5 numeric digits, such as a zip code :
+    zip: ['', [Validators.pattern('^\\d{5}$')]],
+  });
+  contactDetailsFormGroup = this._formBuilder.group({
+    // Matches 99.99% of e-mail addresses (excludes IP e-mails)
+    email: ['', Validators.pattern('^[\\w-]+(?:\\.[\\w-]+)*@(?:[\\w-]+\\.)+[a-zA-Z]{2,7}$')],
+    phone: ['', Validators.required],
+    partnerLastName: [''],
+    partnerFirstName: [''],
+    partnerPhone: ['']
+  });
+  stepperOrientation: Observable<StepperOrientation>;
 
+  constructor(
+    private _formBuilder: FormBuilder,
+    breakpointObserver: BreakpointObserver,
+  ) {
+    this.stepperOrientation = breakpointObserver
+      .observe('(min-width: 800px)')
+      .pipe(map(({matches}) => (matches ? 'horizontal' : 'vertical')));
+  }
+
+
+  ngOnInit(): void {
+  }
+
+  public submit(): void {
+    const identityFormData = this.identityFormGroup.value;
+    const postalAddressFormData = this.postalAddressFormGroup.value;
+    const contactDetailsFormData = this.contactDetailsFormGroup.value;
+    const birthday = identityFormData['birthday'];
+    console.warn(moment(birthday).format('YYYY-MM-DD'));
+  }
 }